Fix Phan after getGuid() return type change
Statement::getGuid() has always been documented to return string|null,
but Phan only started complaining about it once ?string was added as a
static return type (in addition to the existing phpdoc) and strict_types
were declared for Statement.php, which broke our CI.

In SparqlHelper and UpdateConstraintsTableJob, just add a @phan-var
annotation to tell Phan that we know the statement will always have a
nonnull GUID. For the three Context classes, add a small helper function
to AbstractContext: I think this is nice anyways (so the getCursor()
implementations don’t need to introduce any local variables), but more
importantly, hundreds of tests use NewStatement to create test
statements without a GUID, and it’s much easier to create a fake GUID
here than to add it to all those tests. (In CI, 330 tests failed; out of
the ones I looked at, it seemed like most of the tests would need to be
individually fixed otherwise, i.e. I didn’t see an example of dozens of
tests using the same test statement.)

Bug: T317101
Change-Id: Iaaac1925f6a2c480e36580d5c2c8217628b12d7a