Move MessageLocalizer in ViolationMessageRendererFactory
This depends on the user language, so it should be passed into the
factory at the getViolationMessageRenderer() call, together with the
user language as a Language instance, not in the constructor. This
change is a requirement for change I31ee8c21d2.

Also rename some variables to make it clear that the language is
specifically the user language.

Bug: T316936
Change-Id: Ic22fc213cfe8cb6e5baf3414454484c573b5029a