Set explicit targets of ResourceLoader modules
Limit the modules that weren’t already desktop+mobile to desktop-only,
following the rationale at T326405.

For the .gadget module, it probably doesn’t make sense to enable it
while statement editing is desktop-only: if users aren’t able to edit
the statements to fix any violations, we shouldn’t show them the
violations to begin with. The .suggestions module similarly doesn’t make
sense to add while statements can’t be edited on mobile. On the other
hand, the SpecialConstraintReportPage module might well work on mobile,
and is a good candidate for T326428.

Note that the .ui module is already desktop+mobile (since T262489), as
it’s used by WikibaseMediaInfo. This doesn’t contradict the rationale
above: if WikibaseMediaInfo supports editing statements on mobile, then
it also makes sense to show constraint violations on mobile there.

The .icon module is left untouched – ResourceLoaderImageModule seems to
be desktop+mobile by default, and we even get a CI error if we try to
set any targets in it explicitly (fails JSON schema validation).

Bug: T326405
Change-Id: Ia9d466833a101f15bcb347108f9debf5b393cf18