Remove unused @throws from DelegatingConstraintChecker
This was added in change I1771cf4449 (commit 367f502e23) for no reason
that I’ve been able to figure out (nothing else in that change seems
related to InvalidArgumentException). While it’s probably possible for
this method to throw InvalidArgumentException via some of the code it
calls (it calls into every possible constraint checker), I don’t think
this is useful information to include in the phpdoc comment without
further clarification.

Bug: T219037
Change-Id: I74c25185b0c0e217a6ea09426c0b40d5fa7393da