Add strict types to files about to be touched
In DelegatingConstraintChecker, also remove some `= null` default
parameters for methods that are always called with all parameters, i.e.
where `= null` was only used to mark the type as nullable (before proper
nullable types, ?type, were available).

In CheckResult, $constraint needs to keep its doc comment, otherwise
phpcs will complain that the actually unrelated comment “when adding
another status” above it should be a /** doc comment */.

In SpecialConstraintReport::buildExpandableElement(), the documentation
for $expandableContent was actually wrong and null is also possible (and
may be returned by formatParameters()); also, fix a reference to the
variable name in a string that was missed when the code was initially
copy+pasted in change I3b337b347e (commit 786d4e67f4).

Bug: T219037
Change-Id: Iff790c4e19c732f3b438ca4c5127236e3b6ebcac