Don’t create new check result in downgradeResultStatus()
Instead of overwriting the old result with a fresh one, only update the
status and violation message. There’s no need to clear the parsed
constraint parameters, I think, and more importantly we want the
metadata to be preserved, so that we know when the check result is
outdated and needs to be updated. Add a test for this.

This also lets us remove a use of PHP references, which I think is
generally a good thing to do. Also, $context becomes unused, so remove
that as well.

Bug: T219037
Change-Id: Ia259e7efe205e08d5ff747d7c81c368357ea6b51