Show constraint clarifications below violation message
Add a MultilingualTextValue representing the constraint clarification to
CheckResult. (Don’t add it to the constructor, because usually, it’s not
added via the constructor – individual constraint checkers shouldn’t
care about it. Just use the empty value as the default.) Add
serialization and deserialization for caching.

Teach the ConstraintParameterParser to parse constraint clarifications
just like syntax clarifications. This, along with its tests, is really
just copy+pasted with s/syntax/constraint/. Then, make the
DelegatingConstraintChecker parse the constraint clarification after a
successful constraint check (but not if the constraint is not in scope).

Make the wbcheckconstraints API return the constraint clarification
(filtered to the user language, with language fallbacks applied) if it
exists, and teach the gadget to show it if it’s included in the API
response. This should be compatible in both directions: the worst case
is that the constraint clarification won’t be shown if either end
doesn’t understand it.

In both the PHP code and the API response, the constraint clarification
is associated with the check result, not with the constraint being
checked. I think it makes more sense this way: while you could argue
that semantically the clarification is a part of the constraint, the
current representation of a Constraint in PHP and a 'constraint' in the
API output doesn’t contain any *parsed* representation of the constraint
parameters (only the constraint ID, property ID, constraint type item ID
– which, unlike the parameters, is a main snak, not a qualifier – and,
in PHP only, the *unparsed* constraint parameters); since parsing the
constraint parameters, including the constraint clarification, can fail
(e.g. if more than one clarification in the same language was given),
this information and the potential failure is associated with a check
result instead.

Bug: T219037
Change-Id: I3af287f8cc9e78522d77543e21021df3a52c7114