Remove targets definitions
This is done to make CI green again due to the failing
ResourcesTest::testRedundantTargets for Wikibase.

The gadget and suggestions modules are already behind a (not-great)
check for the mobile view.

However, loading the SpecialConstraintReportPage js will enable some minor
folding code on that special page. While that might not be ideal for
touch-control, it seems acceptable.

Depends-On: I92c297f0d170db17dfce03ed0d26614ac2182c87
Needed-By: I0af85955adbb224827bf52e8d4c49d0aca0caba4
Change-Id: I0f53fbda59f6d2e3088aa3f59a0ba795194240a0