Remove constraint parameters from SpecialConstraintReport
The parameters are only rendered on Special:ConstraintReport; given that
almost nobody uses the special page, that’s a lot of extra code for very
little gain, so we’re going to remove it now. Start by removing the
rendered parameters from the special page; the other now-dead code will
be removed in successive commits.

Bug: T338593
Change-Id: I34841c8cf4690584786b28ce8d95566d56dca7a5