Make suggestions API requests sequentially
Make sure that we don’t reorder the results by just making all the
requests in series rather than in parallel, which is also nicer for the
API backend. It does mean the suggestions might load slightly slower,
though. (If that’s a problem, we can look into T220956.)

To avoid having to stub out even more of $.Deferred() in the tests, I
decided to just rewrite the method using only native Promise – I think
the code is still fairly readable. ESLint is fine with that as long as
we remove the browserslist from package.json; recall that IE11 is no
longer supported and was dropped from browserslist-config-wikimedia in
v0.5.0 [1].

[1]: https://gitlab.wikimedia.org/repos/ci-tools/browserslist-config-wikimedia/-/blob/main/CHANGELOG.md#v050-2023-04-12

Bug: T355697
Change-Id: Iccad56f4f2d129b0d758697d7dc08fb4932ebbe3