SpecialConstraintReport: Handle invalid redirect
Try to parse the entity ID and only issue the redirect if it’s valid.
This is *slightly* wasteful (in that the ID is now parsed twice, in the
original request and in the redirect target) but better than throwing
uncaught exceptions into the production error log.

The control flow in the error case is a bit ugly – we can’t directly
show the error message in the catch block (unless we copy+paste the rest
of the code that sets up the special page), so fall through and later
use the presence of the $postRequest as a marker that the error should
be shown. But I think it’s okay, and I can’t think of anything better at
the moment.

Bug: T379820
Change-Id: I28ef26b7b189861110b6b6872a28b63f297256b4