Silence preg_match warnings on Shellbox
`preg_match` returns false, and emits an `E_WARNING` when the regex
pattern is invalid. Shellbox currently turns all emitted errors
and warnings into ShellboxErrors. This is unnecessary for this
check, given that we expect some regexes to be invalid.

In order to silence the warnings, the `preg_match` call needs to be
wrapped in another function, which is sent to Shellbox. Additionally,
Shellbox itself needed a change to its error handler in order to
actually ignore silenced warnings.

Invalid regex should no longer trigger a ShellboxError once this and
the depended-on commit are deployed. If a ShellboxError does occur
after this, the cause cannot be inferred, and therefore should no
longer trigger a `ConstraintParameterException`. Since that inference
is removed, this commit shouldn't be merged until the shellbox commit
has been deployed.

Bug: T380751
Depends-On: I7ef612b94aeb1dd33db0e281be5e66f43f409560
Change-Id: Ib8aaf5a3bf261f114b9e1c5f85f3e35252f908ef