build: Disable MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment sniff
There are 26 files with 67 findings.

This extension has many comments after code lines.
Sniffs listed in <exclude> get autofixed by libup and indicate that the
sniff should be fixed in the long term.
Disable the sniff with serverity = 0 to indicate that
this sniff does not need fixing in this extension.

Change-Id: I3034867c3514caea9f58244602326f1db2e0c365