Declare compatibility with data-values/number 0.13.0
This is needed so we can start to use the new version in production.

Some of the older compatibility declarations in here are probably quite
outdated and simply not true at this point (are we really still
compatible with data-values/data-values 1.0.0, released ten years ago?),
but that should be fixed separately.

Bug: T379481
Change-Id: Iac57b15387a4457c94a1e3fa72468af1ec6d4e66