Replace distinct-values SPARQL queries
The previous queries relied on the data for the current entity being
available in the query service. With the WDQS graph split, this is no
longer guaranteed, so we need new queries that work without this.
Instead of finding the value we’re looking for via triples based on the
original entity / statement, encode the value directly into the query;
we do this by reusing Wikibase’s RDF export code for the Turtle format
(which is close enough to SPARQL to be compatible for this limited
purpose). Then we can easily find other entities with the same value.

The separator code also needs to be rewritten to account for this.
Instead of checking in the SPARQL query which separators the original
statement has and excluding other results based on that, we need to
check in PHP which separators are present, and modify the query based on
that. The logic here gets a bit confusing with the number of negations,
but I think it should be correct.

It’s perhaps worth pointing out that the code for checking this
constraint type on qualifiers and references is pretty damn close to
dead code – the constraint is not checked there by default, and on
Wikidata there is only one property with such a constraint that is
checked on qualifiers [1], and none at all that are checked on
references [2]. (This sounds sensible to me, as I struggle to imagine
any valid use case for this constraint type on qualifiers and
references.) Nevertheless, we support it – the main-snak case is the
more complicated one anyway (qualifiers and references don’t have to
deal with the “separator” parameter), so supporting qualifiers and
references bears very little additional cost.

[1]: https://w.wiki/CkdX
[2]: https://w.wiki/CkdZ

Bug: T369079
Change-Id: I72365edf417d35a72326b77b68f98a021860948e
Depends-On: Iadb3f55587aad16836ae965a53a43dec47c8ce14
(cherry picked from commit 9d502a8e30508773ec59259d8cae13f26ab1efe0)