Improve SPARQL query construction in SparqlHelper
Previously, we partially used RdfVocabulary, but also hard-coded some of
the namespaces, even though they may be configurable. Try to use
RdfVocabulary more consistently. This should make no difference in
Wikimedia production (even on Commons – the prefix for items there is
still wd:, not something else), but may help other installs.

I’ve only updated the uses related to hasType(). The various functions
related to distinct-values / unique-values constraints are basically
rewritten entirely in the following change I72365edf41 anyways, so I
don’t think it’s worth spending time on cleaning up the old code that’s
about to be removed. (But I wanted to introduce this pattern and use it
for hasType() in a separate change, because that change is already big
enough as it is.)

To be able to call RdfVocabulary::getEntityRepositoryName(), we need to
have the entity IDs as EntityId objects rather than strings. For the
$id, this is easily done: TypeCheckerHelper already has it as an ID, we
just need to stop turning into a string. On the other hand, the $classes
are always an array of strings from the moment they come out of the
ConstraintParameterParser, and they’re also passed around in several
other places, so I don’t really want to turn those into ItemId objects
everywhere; calling the ItemId constructor in the $classesValues
array_map() callback is good enough, IMHO.

Bug:T369079
Change-Id: Icfd242feacefe320f5e1be2d7999746bcc10f79c
(cherry picked from commit 282a591953f6a9257ea3d4152aecc7a82ab1fa50)