Somewhat improve RdfVocabulary usage in SparqlHelper
The old code, directly calling getNamespaceURI() with
RdfVocabulary::NS_STATEMENT and RdfVocabulary::NS_VALUE, implicitly made
two assumptions: 1) that these two namespaces use the predicate prefix
rather than the node prefix, and 2) that the predicate prefix is the
empty string. (For more correct usage of getNamespaceURI() without such
assumptions, see the other calls below the code touched here, where it’s
called with names from $rdfVocabulary->propertyNamespaceNames.)
Assumption 1 was correct as far as the existing Wikibase code went, but
really represents a bug in Wikibase that we want to fix in T384344;
assumption 2 was always wrong to make: it’s true for Wikidata, and for a
default Wikibase without customized entity sources, but it’s not always
true in general.

As part of T384344, Wikibase will change these namespaces from the
predicate prefix to the node prefix (on Wikidata: from "" to "wd"). This
causes the old calls in this code to fail with an OutOfBoundsException
because the namespaces with the empty string as the prefix are no longer
defined. Fix that by getting the namespace name from the RdfVocabulary
instead, based on the entity source for items (which will be the correct
entity source for this context on Wikidata and Commons), which should
work both before and after the Wikibase change. Note that we still
hard-code the prefixes we write into the SPARQL query ("wds" and "wdv" –
i.e. these always used the node prefix, rather than the predicate
prefix, and assumed that it was "wd"!): there is currently a mismatch
between Wikibase and WDQS which this hard-coding works around (and has
always worked around, without us realizing it), and until that’s fully
resolved by T384344, we can’t get the namespace name from the
RdfVocabulary yet.

In the tests, this requires us to flesh out the RdfVocabulary that we
inject into the SparqlHelper, so that it will know the namespaces we now
want from it.

The usage is further improved in changes Icfd242feac and I72365edf41.

Bug: T374021
Bug: T384344
Change-Id: I7a1b829209e772d710f2e1e598cb660542ffbf9e