Introduce rights for checking constraints
This adds two new rights: wbqc-check-constraints for checking
constraints (and constraint parameters) via the API, and
wbqc-check-constraints-uncached for checking constraints via the special
page. Because the special page bypasses the cache, it is potentially
much more expensive, so that right is only given to logged-in users by
default: anonymous visitors are no longer allowed to use the special
page. Of course, both permissions can be reassigned by wiki admins as
desired.

We also check the new permission to determine whether the constraints
gadget should be loaded in the first place. I think this could make the
$wgWBQualityConstraintsShowConstraintViolationToNonLoggedInUsers config
obsolete – if you don’t want to show constraint violations to anonymous
visitors, simply remove the right from them:

    $wgGroupPermissions['*']['wbqc-check-constraints'] = false;
    $wgGroupPermissions['user']['wbqc-check-constraints'] = true;

That said, I’m not removing the config yet.

Bug: T401789
Change-Id: I05b58a02ff800c77b9580e4bb3523bd38a8835ee