Remove config to show constraint violations to everyone
According to the comments in T272132, this was in fact always meant to
be a temporary setting (T272132#8450103: “we can remove the setting
again”); if there was ever a decision to keep it permanently (I feel
like I remember something but I’m not sure), it doesn’t seem to have
been recorded in any permanent place.

See the parent change (I05b58a02ff) for how to emulate the effect of
this config setting using the new wbqc-check-constraints right.

Bug: T272132
Bug: T401789
Change-Id: I16700b77bc31bb040d748daa6cc1e30ea8ba87a7