Basic tests of QualityConstraints UI on Desktop
This changes the item to have a string statement and stubs out a
wbcheckconstraints response for it, then makes some assertions on the
resulting constraint report popup.

This initial patch only tests the reports for the mainsnak, and only
in the desktop view. The desktop implementation is already covered by
Jasmine tests. For wbui2025 / MEX, we also expect to include Jest
tests of Vue components for the constraint check reports and only to
use Cypress for checking that the end-to-end (API response to UI)
integration works.

The suite may later be extended to implement regression tests for
issues that escape unit testing into production.

Bug: T411606
Change-Id: I97cb6fcc92d8066555a3b3bc531eef0d50088610