Enrich gadget.js and gadget.spec.js with modern JS
Instead of a “poor man’s Object.assign()”, use the real Object.assign(),
which is now available in every environment we support; and instead of a
“poor man’s Promise.all()”, use proper async/await syntax. (This also
makes us send the wbcheckconstraints requests in sequence, rather than
in parallel, which is more in line with the API etiquette [1].)

In my opinion, the gadget code is now much nicer than before. The test
code is a bit more of a mixed bag – the way it worked was pretty
bizarre, but it successfully made the asynchronous-looking code in
gadget.js work fully synchronously, so that the test could immediately
run assertions on it; with async/await that’s no longer an option, so
the tests have to await something (in one case, where we have nothing
concrete to await, we just wait for the next microtick). I’m not sure to
what extent that’s an improvement.

[1]: https://www.mediawiki.org/wiki/API:Etiquette#Request_limit

Change-Id: Ieacf2259cf16cab5b5a8a7321291e9fa2813fe33