Introduce Cypress tests to WikibaseQualityConstraints
For the MEX / wbui2025 implementation of Quality Constraints, we want
to be able to run UI / browser tests.

Install cypress, cypress-axe and cypress-wikibase-api. Per the change
I6a6a69642926cbe6, we need to use an older version of api-testing
because of an issue with one of the shared dependencies between
cypress and api-testing.

The correct configuration of WikibaseQualityConstraints in CI would
be complicated and require us to be able to run maintenance scripts
as part of the CI setup (which we do not currently have a mechanism
for). As a result, the browser tests will need to stub the calls to
the `wbcheckconstraints` API so make UI elements appear. We assume
that the API is stable, and that the tests will be updated as the
API changes.

Bug: T411606
Change-Id: I7c86313e7dd286e0859bf820b935f0a718aafc30