Fetch and store constraint violation information
Refactor the existing gadget code to make it possible to create a
single "gadget" that supports both the mobile and desktop UIs. The
behaviour of the gadget is slightly different depending on which UI
is in use - we preserve the behaviour of the desktop gadget whilst
also using the 'Indicator' support recently introduced to the wbui2025
implementation.

The code for wbui2025 relies on building HTML blobs using JS template
strings. This is less a maintainability and security concern, though
we hope that here (and in the review of the now-abandoned change
I76c9ceac3ee3 from which this change is derived) that we have caught
all potential escaping / XSS issues in the current implementation.

Bug: T411602
Change-Id: I6eb4d26c07b9a73405dedcf08905f5c0a36c6799